fix(ui-auth): bounce to login when the live stream closes on an expired session
